+++ functions/Send-GraphiteMetric.ps1
@@ -57,7 +57,13 @@
             } elseif ($env:GRAPHITE_ACCESS_TOKEN) {
                 $env:GRAPHITE_ACCESS_TOKEN
             }
-        )
+        ),
+
+        # Switch to output result to the console.
+        [switch] $OutputToConsole,
+
+        # Switch to still return the reponse even when the '-OutputToConsole' parameter is provided.
+        [switch] $PassThru
     )
 
     begin {
@@ -115,7 +121,19 @@ Exception:
             return
         }
 
+        # Output to console
+        if ($OutputToConsole.IsPresent) {
+            $columns = @(
+                @{ label = 'Status'; expression = { "$($_.StatusCode) $($_.StatusDescription)" } }
+                @{ label = 'Published'; expression = { "$(($_.Content | ConvertFrom-Json).Published)" } }
+                @{ label = 'Invalid'; expression = { "$(($_.Content | ConvertFrom-Json).Invalid)" } }
+            )
+            $response | Select-Object $columns | ForEach-Object { $_ | Out-Host }
+        }
+
         # Output response
-        $response
+        if (-Not $OutputToConsole.IsPresent -Or $PassThru.IsPresent) {
+            $response
+        }
     }
 }